# api-docs.yaml
openapi: 3.1.0
info:
  title: WMU Developer Club's Financial API
  description: An API to interact with the backend of WMU Developer Club's database
    to manage finances and budgeting.
  contact:
    name: API Support Technician
    email: bluewheatthin@gmail.com
  version: v1.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /admin/groups/search:
    put:
      tags:
      - Finance Groups
      summary: Filters through finance groups based on specified values
      description: "Takes in a JSON array, where each element is a Filter object consisting\
        \ of the column to filter by, the operation to filter based on, and the desired\
        \ value, and returns all of the rows in the Finance Group table which match\
        \ the Filter objects"
      operationId: filterFinanceGroups
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /admin/groups/remove_group/id={id}:
    put:
      tags:
      - Finance Groups
      summary: Removes a finance group
      description: Takes in the id of a Finance Group and sets the deleted column
        to 1. Returns the FinanceGroup object on success
      operationId: removeGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /admin/groups/create_group:
    post:
      tags:
      - Finance Groups
      summary: Adds a new finance group
      description: Takes in a JSON representation of a FinanceGroup object and adds
        it to the Finance Group table. Returns the FinanceGroup object on success
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinanceGroup"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    Filter:
      type: object
      properties:
        col:
          type: string
          description: The column to filter by
          example: feeFlag
        op:
          type: string
          description: The operator to filter by
          example: leq
        val:
          type: object
          description: The value to filter by
          example: 5
      required:
      - col
      - op
      - val
...
    FinanceGroup:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
          description: The title of the finance group
          example: Admin
        deleted:
          type: integer
          format: int32
          description: A flag for if this finance group is deleted or not
          example: 1
      required:
      - deleted
      - title
...